block,elevator: use new hashtable implementation
Switch elevator to use the new hashtable implementation. This reduces
the amount of generic unrelated code in the elevator.

This also removes the dymanic allocation of the hash table. The size of
the table is constant so there's no point in paying the price of an
extra dereference when accessing it.

Signed-off-by: Sasha Levin <sasha.levin@oracle.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>

Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>